Task 69657: Hangfire extensions: Prověřit spouštění úloh při startu aplikace - záměrem bylo v debugu zabránit jejich spuštění

+++ Havit.Hangfire.Extensions/BackgroundJobs/BackgroundJobManager.cs
@@ -28,18 +28,15 @@ public class BackgroundJobManager : IBackgroundJobManager
 		// remove enqueued jobs
 		List<string> toDelete = new List<string>();
 		IMonitoringApi monitor = jobStorage.GetMonitoringApi();
-		if (monitor.Queues().Any())
-		{
-			QueueWithTopEnqueuedJobsDto queueWithTopEnqueuedJobs = monitor.Queues().SingleOrDefault(q => q.Name == queue); // get the single queue by name
+		QueueWithTopEnqueuedJobsDto queueWithTopEnqueuedJobs = monitor.Queues().SingleOrDefault(q => q.Name == queue); // get the single queue by name
 
-			if (queueWithTopEnqueuedJobs != null)
-			{
-				Enumerable.Range(0, (int)Math.Ceiling(queue.Length / 1000d)) // lets batch items by 1000
-					.SelectMany(batchIndex => monitor.EnqueuedJobs(queueWithTopEnqueuedJobs.Name, 1000 * batchIndex, 1000)) // select jobs in a batch by batchIndex
-					.Select(jobEntry => jobEntry.Key) // select JobId (key is a JobId)
-					.ToList() // process all pages to memory
-					.ForEach(jobId => backgroundJobClient.Delete(jobId)); // remove the job
-			}
+		if ((queueWithTopEnqueuedJobs != null) && queueWithTopEnqueuedJobs.FirstJobs.Any())
+		{
+			Enumerable.Range(0, (int)Math.Ceiling(queueWithTopEnqueuedJobs.Length / 1000d)) // lets batch items by 1000
+				.SelectMany(batchIndex => monitor.EnqueuedJobs(queueWithTopEnqueuedJobs.Name, 1000 * batchIndex, 1000)) // select jobs in a batch by batchIndex
+				.Select(jobEntry => jobEntry.Key) // select JobId (key is a JobId)
+				.ToList() // process all pages to memory
+				.ForEach(jobId => backgroundJobClient.Delete(jobId)); // remove the job
 		}
 	}
 }
